Replaced the latex-tools-specific option with general information.
It looks like that the '% !TEX option' is a latex-tools-specific statement, which doesn't help e.g. the "latex" package of Atom. Therefore, I replaced it with more information about how to configure it. 

Although this information now exists in the README.md, I think it is good to also have it in the main.tex to ensure that the user will see it.

Closes #5.