Removed the (obsolete) shortoverview.sty.
No usage example is provided in this or in the previous template.
The package looks custom, as it is written with german comments, it has no documentation and it doesn't appear in online searches in different contexes than this template.

I guess that the purpose was to include an overview of the contents of each chapter in the begining of it. If desired, we could find a more modern way to do that.