Removed the Babel dependency (not used).
The package Babel is one of the available ways to write documents with multiple languages in *TeX.

In this template, the package was configured to only use english, so it was not actually used. The suggested way to write documents with multiple languages now is to use another engine, e.g. XeLaTeX.

Currently, the template compiles successfully with pdflatex (LaTeX), XeLaTeX and LuaLaTeX.