Comment clean-up and updates in main.tex.
I included all the files that are currently provided in the template and updated the comments.

I changed the tabs to spaces in some areas because in those areas, tabs are only causing troubles. Compare the new and the old versions in different tab sizes.

I removed some lonely comments before the input commands, because they were not used everywhere and they didn't provide any new information.

Last, I removed the extension from the `\input{components/text.tex}` because it is not needed and it is inconsistent with the rest of the document.