Improved the hyperref color handling.
Using the package `ocgx2`, the colors of the links are disabled when printing. This doesn't pose the  same limitations in text wrapping as when providing the `ocgcolorlinks` option to `hyperref`.

Note that in order to see the result, a compatible PDF viewer is required. Currently, this part of the PDF standard is not implemented in e.g. Evince, Okular. It works though in Adobe Reader, Foxit Reader, in the embeded PDF reader of Google Chrome/Chromium and other. If it is not supported, the links simply remain colored.

A different color was assigned to each type of hyperlink (internal hyperlinks - blue, web links - light blue, citations - orange). Colors from the TUM Corporate Design guide were chosen and can be also used in other parts of the document. The user can change them and even set them to black to hide them, if the `ocgx2` solution is not enough. This closes #13.

The package `color` was replaced by the package `xcolor`, as it provides more options (e.g. setting hex-colors). All the color-related options were moved to a new section "Color".